+++ NEWS.md
@@ -1,8 +1,18 @@
+# usemodels 0.2.0
+
+* SVM (#14) and C5.0 (#9) models were added. 
+
+* Clipboard access is enabled in all `use_*()` functions using the `clipboard = TRUE` argument (#12).
+
+* Use new `all_nominal_predictors()` selector (#17).
+
 # usemodels 0.1.0
 
 * Updated version requirements for dependencies. 
 
 * Added `use_cubist()`. 
+
+* Fixed bug in handling of variables for `step_string2factor()`
 
 # usemodels 0.0.1